clickhouse/requirements.txt
requests==2.32.5
python-decouple==3.8
